# doc/weloveapps_wert-swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "title" : "SWE",
    "version" : "master"
  },
  "host" : "swe-test.wla-backend.de",
  "basePath" : "",
  "schemes" : [
    "https"
  ],
  "paths" : {
...
    "/benefits/{id}" : {
      "get" : {
        "summary" : "details",
        "description" : "Get details for a single benefit, identified by `id`.",
        "operationId" : "details",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "description" : "Event Id",
            "required" : true,
            "type" : "string"
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object",
              "properties" : {
                "result" : {
                  "type" : "object",
                  "properties" : {
                    "contact" : {
                      "type" : "object",
                      "properties" : {
                        "email" : {
                          "type" : "string",
                          "example" : "heimvorteil@stadtwerke-erfurt.de"
                        }
                      }
                    },
                    "_id" : {
                      "type" : "string",
                      "example" : "bfd5fc22-ab37-4b47-b242-2cebe15d59da"
                    },
                    "title" : {
                      "type" : "string",
                      "example" : "MEIN HEIMVORTEIL"
                    },
                    "subtitle" : {
                      "type" : "string",
                      "example" : "Lichterfest im egapark!"
                    },
                    "date_start" : {
                      "type" : "string",
                      "example" : "2018-06-14T22:00:00.000Z"
                    },
                    "date_end" : {
                      "type" : "string",
                      "example" : "2018-07-14T21:59:00.000Z"
                    },
                    "description" : {
                      "type" : "string",
                      "example" : "Am 10. und 11. August 2018 geht das Lichterfest in seine 52. Auflage. Der Saisonhöhepunkt im größten Garten Thüringens sorgt Jahr für Jahr für abertausende bunte Lichter, Musik auf mehreren Bühnen, Laser-, Licht- und Feuershows. Und was auf keinen Fall fehlen darf, ist ein spektakuläres Feuerwerk an beiden Abenden. Für das diesjährige Lichterfest verlosen wir an SWE Energiekunden 50 x 2 Freikarten. Als Teilnehmer am Gewinnspiel erklären Sie sich verbindlich, dass Sie die Teilnahmebedingungen gelesen und akzeptiert haben.\n\nWir wünschen Ihnen viel Spaß und einen schönen Sommer!\n\nIhr Team der SWE Energie GmbH"
                    },
                    "image" : {
                      "type" : "string",
                      "example" : "https://www.stadtwerke-erfurt.de/pb/site/swe/get/params_E2084650003/273126/Header%20HV%20Lichterfest%202018%20635x220.jpg"
                    },
                    "source" : {
                      "type" : "string",
                      "example" : "swe-energie"
                    },
                    "updated_at" : {
                      "type" : "string",
                      "example" : "2018-06-18T07:55:00.000Z"
                    },
                    "annotation_title" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          },
          "not-found" : {
            "$ref" : "#/responses/not-found"
          },
          "unprocessable-entity" : {
            "$ref" : "#/responses/unprocessable-entity"
          },
          "internal-server-error" : {
            "$ref" : "#/responses/internal-server-error"
          }
        },
        "x-speca-id" : "details"
      }
    },
    "/benefits" : {
      "get" : {
        "summary" : "list",
        "description" : "Retrieves a list of benefits. Sorted by ascending start date of the benefits. If the `limit` parameter is not set, 100 events are returned by default.",
        "operationId" : "list",
        "parameters" : [
          {
            "$ref" : "#/parameters/skip"
          },
          {
            "$ref" : "#/parameters/limit"
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/benefit list schema"
            }
          },
          "unprocessable-entity" : {
            "$ref" : "#/responses/unprocessable-entity"
          },
          "internal-server-error" : {
            "$ref" : "#/responses/internal-server-error"
          }
        },
        "x-speca-id" : "list"
      }
    },
...
    "/benefits/source/{source}" : {
      "get" : {
        "summary" : "list for source",
        "description" : "Retrieves a list of benefits for the given source. Sorted by ascending start date of the benefits. If the `limit` parameter is not set, 100 benefits are returned by default.\n\nValid source identifiers are:\n- swe-energie\n- thuewa\n- swe-netz\n- swe-stadtwirtschaft\n- swe-parken\n- evag\n- swe-baeder\n- ega\n- buga",
        "operationId" : "list-for-source",
        "parameters" : [
          {
            "$ref" : "#/parameters/skip"
          },
          {
            "$ref" : "#/parameters/limit"
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/benefit list schema"
            }
          },
          "unprocessable-entity" : {
            "$ref" : "#/responses/unprocessable-entity"
          },
          "internal-server-error" : {
            "$ref" : "#/responses/internal-server-error"
          }
        },
        "x-speca-id" : "list-for-source"
      }
    },
    "/benefits/{id}/fave" : {
      "post" : {
        "summary" : "fave",
        "description" : "Fave a benefit specified by `id`. Adds the benefit id to the list of faved benefits for this user. See [/auth/me](speca://me).",
        "operationId" : "fave",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "description" : "Id of the benefit that will be faved.",
            "required" : true,
            "type" : "string"
          }
        ],
        "responses" : {
          "204" : {
            "description" : "Benefit was successfully faved."
          },
          "not-found" : {
            "$ref" : "#/responses/not-found"
          },
          "internal-server-error" : {
            "$ref" : "#/responses/internal-server-error"
          }
        },
        "security" : [
          {
            "Auth" : [ ]
          }
        ],
        "x-speca-id" : "fave"
      }
    },
    "/benefits/{id}/unfave" : {
      "post" : {
        "summary" : "unfave",
        "description" : "Unfave a benefit specified by `id`. Removes the benefit id from the list of faved benefits for this user. See [/auth/me](speca://me).",
        "operationId" : "unfave",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "description" : "Id of the benefit that will be unfaved",
            "required" : true,
            "type" : "string"
          }
        ],
        "responses" : {
          "204" : {
            "description" : "Benefit was successfully unfaved."
          },
          "not-found" : {
            "$ref" : "#/responses/not-found"
          },
          "internal-server-error" : {
            "$ref" : "#/responses/internal-server-error"
          }
        },
        "security" : [
          {
            "Auth" : [ ]
          }
        ],
        "x-speca-id" : "unfave"
      }
    },
    "/benefits/{id}/redeem" : {
      "post" : {
        "summary" : "redeem",
        "description" : "Redeem a benefit specified by `id`. Adds the benefit id to the list of redeemed benefits for this user and triggers a redeem notification (via email) for SWE. See [/auth/me](speca://me).",
        "operationId" : "redeem",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "description" : "Id of the benefit that should be redeemed",
            "required" : true,
            "type" : "string"
          }
        ],
        "responses" : {
          "204" : {
            "description" : "Benefit was successfully redeemed."
          },
          "not-found" : {
            "$ref" : "#/responses/not-found"
          },
          "409" : {
            "description" : "Benefit was already redeemed.",
            "schema" : {
              "type" : "object",
              "properties" : {
                "result" : {
                  "type" : "object",
                  "properties" : {
                    "message" : {
                      "type" : "string",
                      "example" : "Benefit was already redeemed."
                    }
                  }
                }
              }
            }
          },
          "internal-server-error" : {
            "$ref" : "#/responses/internal-server-error"
          }
        },
        "security" : [
          {
            "Auth" : [ ]
          }
        ],
        "x-speca-id" : "redeem"
      }
    },
...
  },
...
  "parameters" : {
    "skip" : {
      "name" : "skip",
      "in" : "query",
      "description" : "Used for pagination. Skips the first n results. Default value is 0 if no skip is specified.",
      "type" : "string",
      "x-speca-id" : "skip"
    },
    "limit" : {
      "name" : "limit",
      "in" : "query",
      "description" : "Specifies the number of returned events. Maximum is 1000. Default value is 100 if no limit is specified.",
      "type" : "string",
      "x-speca-id" : "limit"
    }
  },
...
  "securityDefinitions" : {
    "Auth" : {
      "type" : "oauth2",
      "flow" : "password",
      "tokenUrl" : "/auth/login"
    }
  }
}